# Conflicts:
#	Dependencies.cmake.in
#	vcpkg.json

+++ lib/src/engine/bof2d_text_renderer.cpp
@@ -139,7 +139,7 @@ void TextRenderer::Load(std::string font, unsigned int fontSize)
         texture,
         glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
         glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-        face->glyph->advance.x
+        (int)face->glyph->advance.x
     };
     Characters.insert(std::pair<char, Character>(c, character));
   }
